libsdl-x11-sudo-issue: sample program to demonstrate libsdl X11 sudo issue
The purpose of this program is to demonstate and find the best
resolution to the libsdl[12] X11 issue of when using SDL with sudo. The
claim is that SDL library already does the best it can so the goal
is to provide an example of something more reasonable for developers
to do in this case.

The respective upstream SDL bug is:

https://bugzilla.libsdl.org/show_bug.cgi?id=2446

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>